Added servo support and stabilized color
Color readings now are unscathed and more stable, any WOODEN BLOCK can be read from a distance of about 5 mm to 6 mm away. Need to add support for the glossy balls in the future, possibly by switching from the internal LED (which is needed to measure wooden blocks) to an external LED. Since the balls are glossy, the LED on the sensor reflects off, and thus an external LED is needed to internally light the balls to avoid false readings.