Code for the old color sensor
This is code for the DFRobot color sensor, which requires more distance between itself and the object, can only report one color at a time, and reports color via the frequency of a wave instead of via analog values which can instantly be sampled. Its longer detection distance may make it better for measuring color under the robot.