Flag geometry test
Now that the SD card won't be used, we need a new way to draw the flag. As it turns out, drawing the flag with geometry isn't as hard as it first seemed, and is in fact faster to draw to the screen than a bitmap. Still working on finishing the flag and the raising animation, but once it is done it can be implemented into the current screen code (screen_master) without issue.