Geometry flag, cleaned code, renamed
People said the name was weird so I replaced the master part with the main. 
More importantly, now the flag is drawn entirely with geometry--it looks exactly the same as the original flag, but can be drawn very quickly (maybe in a 10th of a second?).
I also took the time to go through the code and take out some extra stuff. No need for the SD library any more, or the SPI library which wasn't even used anyway.
I think I've figured out a way to make the flag move: draw the flag with offsets that put it off of the screen, then loop the flag drawing code and reduce the offset amount a little each time to make it seem like the flag is rising up from the bottom of the screen (the origin of the coordinate system is at the top left of the screen).
I also need to implement the serial sending and voltage receiving parts of the code, which I will do once the screen is all set flag-wise.