Code for the nano (NOT THE MICRO)
As of the time of writing, this code is basically identical to the primary code; however, this code scrolls the flag. For some reason the Micro can't draw to the screen as quickly as the nano, so we can't use it to make the flag seem like it is rising. Note that the micro is on the 2nd batch of screen PCBs.