Text has been made vertical, stop button is red now
Make sure you tap the corners of the buttons, the stop button is a little picky about where you pick. The last code worked fine, but the text wasn't vertical--if there are any problems with this code then please revert to the previous version. I made some improvements to the flag image as well.